infrastructure/requirements.txt
# Infrastructure testing and validation dependencies
pytest>=7.0.0
pytest-html>=3.1.0
pytest-cov>=4.0.0
requests>=2.28.0
docker>=6.0.0
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
elasticsearch>=7.15.0
pyyaml>=6.0
jinja2>=3.1.0

# Optional dependencies for advanced testing
paramiko>=2.11.0  # SSH testing
pysmb>=1.2.0     # SMB testing
dnspython>=2.2.0  # DNS testing
scapy>=2.4.0      # Network packet analysis